slacker/requirements.txt
cachetools
colorama
jinja2
requests
pyyaml
